requirements.txt
ipykernel
flask
numpy
pandas
matplotlib
tensorflow
pathlib
tqdm
pillow
albumentations
-e.